Migrate all gallery scripts to ES6 bundle
