+++ assets/css/_gallery.scss
@@ -262,6 +262,10 @@ body:not(.safari).gallery-visible #street-section-sky {
   }
 }
 
+#gallery-try-again {
+  margin-left: 10px;
+}
+
 body:not(.gallery-no-move-transition) #main-screen,
 body:not(.gallery-no-move-transition) #gallery,
 body:not(.gallery-no-move-transition) #street-name-canvas,
