+++ assets/js/app.js
@@ -6,5 +6,4 @@
 
 // = require_tree segments
 // = require_tree streets
-// = require_tree gallery
 // = require_tree users
